Choose ship screen
There's now a screen after title and disconnect that prompts the player
to enter the name of a ship (cheetah, gull, etc.).  The client will send
this as their ship choice then go into a waiting state. If the given
name isn't valid the server will send back badShipError:true, which will
prompt the client to go back to the choose ship screen.