function queue
Each game now has an array of functions that will be called and consumed
after checkCollisions runs. This is mainly so the remoteInput can stamp
objects after their states (particularly hitscans) are finalized for the
frame.

This eliminates the bug with lasers not showing up as shortened after
collisions on the client.