Sparse grid spatial optimization pt1
Added a new array wrapper, SuperArray, which will allow me to avoid a
bit of memory allocation and deallocation. Two super arrays have been
added to game, reportQueue and tileArray.

Objects are added to reportQueue by updaters.queueReport, which is run
for every object. It also checks the object's x and y against
reportQueue.max and reportQueue.min. After updates and before
collisions, processReportQueue will be called on the game, which just
clears the reportQueue and resets the max and min for now.

tileArray is just an empty super array for now, but it will eventually
store information for the grid tiles.