spatial optimization pt 2
Everything is implemented except the function to pull objects from the
tileArray using a position and radius, and there's no formula for
deriving precision yet - it's a flat number. Tiles will be created if
the calculated tileArray size is larger than the current number of
tiles, tiles will be cleared, every item in the report queue is sorted
into a tile using the map functions, then the report queue is cleared
and the tileArray's min and max is reset.