Spatial optimization works
At long last the spatial hashing is finally actually improving
performance. Instead of determining a range of cells to add the object
to, I get the indices of the cells the object center and each of the
corners (min/max) are in and add it to those, ignoring duplicates. This
is (for some reason) faster, but it means an object's bounding box can't
be larger than my precision value. This won't matter for the objects
themselves, but bounding box size is affected by velocity, so this means
the object can't be moving fast enough that its bounding box is larger
than the cell size. Precision is 120,000 at the moment, so I don't know
that this will ever be a relevant constraint.